Splitter opp behandling i flere filer og endrer statuser felles
